code rafcator(create differet buttons for exit,share and download)
